chore: split segment and parsing buf to another file
